// src/shared/types/services/binance.ts
import type { ITokenTickerBaseData } from "../web3";

export interface IBinanceTickerData {
  symbol: string;
  priceChange: string;
  priceChangePercent: string;
  weightedAvgPrice: string;
  prevClosePrice: string;
  lastPrice: string;
  lastQty: string;
  bidPrice: string;
  bidQty: string;
  askPrice: string;
  askQty: string;
  openPrice: string;
  highPrice: string;
  lowPrice: string;
  volume: string;
  quoteVolume: string;
  openTime: number;
  closeTime: number;
  firstId: number;
  lastId: number;
  count: number;
}

export interface IGetTopUSDTokensResponse {
  data: Record<string, ITokenTickerBaseData> | ITokenTickerBaseData[];
}